+++ src/interfaces/dependencies.py
@@ -1,17 +1,10 @@
-from fastapi import Depends
 from src.service.login_service import KakaoLoginService
 from src.service.user_service import UserService
 from src.service.auth_service import AuthService
 from src.service.weather.weather_checker import WeatherChecker
 from src.service.prewalk_orchestrator import PrewalkOrchestrator
 
-"""
-이 파일에서 정의된 서비스 객체를 다른 API 파일에서 전역적으로 사용하시면 됩니다!
-예시:
-     @router.get("/users/me")
-     def get_my_profile(user_service: UserService = Depends(get_user_service)):
-         return user_service.get_profile(...)
-"""
+# 이 파일에서 정의된 서비스 객체를 다른 API 파일에서 전역적으로 사용하시면 됩니다!
 
 # 싱글톤 패턴
 weather_checker = WeatherChecker()
