+++ src/main.py
@@ -2,8 +2,9 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from contextlib import asynccontextmanager
 from src.entity.base import init_db
-from src.api import weather_router, user_router, prewalk_router, auth_router, login_router
+from src.interfaces.api import weather_router
 
+from src.interfaces.api import auth_router, login_router, prewalk_router, user_router
 import uvicorn
 
 app = FastAPI()
